1 - Finished search method (fully functional but will improve on it later to detect ".com", ."net" etc...) 2 - Updated hashmap's 2nd parameter to String[] 3 - Looped the home menu and added exit
Note : I don't know if y'all were having the same issue but intellij defaulted to jdk 11 on mine. I had to download and set up jdk 13 and then mess a bit with the source folder to run it on my computer. I think I figured it out so shoot me a message if you're having trouble setting it up